Refactoring ...
