+++ companion/src/firmwares/opentx/opentxinterface.cpp
@@ -24,7 +24,6 @@
 #include "opentxeeprom.h"
 #include "rlefile.h"
 #include "appdata.h"
-#include "storage_sdcard.h"
 
 #define OPENTX_FIRMWARE_DOWNLOADS        "http://downloads-22.open-tx.org/firmware"
 #define OPENTX_NIGHT_FIRMWARE_DOWNLOADS  "http://downloads-22.open-tx.org/nightly/firmware"
@@ -171,6 +170,9 @@ bool OpenTxEepromInterface::loadModelFromRLE(ModelData & model, RleFile * rleFil
 template <class T, class M>
 bool OpenTxEepromInterface::saveToByteArray(const T & src, QByteArray & data, uint8_t version)
 {
+  if (version == 0) {
+    version = getLastDataVersion(getBoard());
+  }
   QByteArray raw;
   M manager((T&)src, board, version, 0);
   // manager.Dump();
@@ -257,97 +259,6 @@ void registerOpenTxEEpromInterfaces()
   registerOpenTxEEpromInterface(BOARD_TARANIS_X9DP);
   registerOpenTxEEpromInterface(BOARD_TARANIS_X9E);
   registerOpenTxEEpromInterface(BOARD_HORUS);
-}
-
-template <class T, class M>
-bool loadFromByteArray(T & dest, const QByteArray & data)
-{
-  foreach(OpenTxEepromInterface * eepromInterface, opentxEEpromInterfaces) {
-    if (eepromInterface->loadFromByteArray<T, M>(dest, data)) {
-      return true;
-    }
-  }
-
-  return false;
-};
-
-bool loadFile(RadioData & radioData, const QString & filename)
-{
-  StorageSdcard storage;
-
-  storage.read(filename);
-
-  // Radio settings
-  qDebug() << "Radio settings:" << storage.radio.size();
-  if (!loadFromByteArray<GeneralSettings, OpenTxGeneralData>(radioData.generalSettings, storage.radio))
-    return false;
-
-  // Models
-  int modelIndex = 0;
-  QString modelList = QString(storage.modelList);
-  QList<QByteArray> lines = storage.modelList.split('\n');
-  int categoryIndex = -1;
-  foreach (const QByteArray & line, lines) {
-    if (!line.isEmpty()) {
-      if (line.startsWith('[') && line.endsWith(']')) {
-        QString name = line.mid(1, line.size() - 2);
-        CategoryData category;
-        strncpy(category.name, name.toStdString().c_str(), sizeof(CategoryData::name));
-        radioData.categories.push_back(category);
-        categoryIndex++;
-      }
-      else {
-        qDebug() << "Loading" << line;
-        foreach (const ModelFile &model, storage.models) {
-          if (line == model.filename) {
-            if (!loadFromByteArray<ModelData, OpenTxModelData>(radioData.models[modelIndex], model.data))
-              return false;
-            strncpy(radioData.models[modelIndex].filename, line.data(), sizeof(radioData.models[modelIndex].filename));
-            radioData.models[modelIndex].category = categoryIndex;
-            radioData.models[modelIndex].used = true;
-            modelIndex++;
-          }
-        }
-      }
-    }
-  }
-
-  return true;
-}
-
-int OpenTxEepromInterface::saveFile(const RadioData & radioData, const QString & filename)
-{
-  StorageSdcard storage;
-  uint8_t version = getLastDataVersion(board);
-
-  // models.txt
-  storage.modelList = QByteArray();
-  int currentCategoryIndex = -1;
-
-  // radio.bin
-  saveToByteArray<GeneralSettings, OpenTxGeneralData>(radioData.generalSettings, storage.radio, version);
-
-  // all models
-  for (unsigned int i=0; i<radioData.models.size(); i++) {
-    const ModelData & model = radioData.models[i];
-    if (!model.isEmpty()) {
-      QString modelFilename = model.filename;
-      QByteArray modelData;
-      saveToByteArray<ModelData, OpenTxModelData>(model, modelData, version);
-      ModelFile modelFile = { modelFilename, modelData };
-      storage.models.append(modelFile);
-      int categoryIndex = model.category;
-      if (currentCategoryIndex != categoryIndex) {
-        storage.modelList.append(QString().sprintf("[%s]\n", radioData.categories[model.category].name));
-        currentCategoryIndex = categoryIndex;
-      }
-      storage.modelList.append(modelFilename + "\n");
-    }
-  }
-
-  storage.write(filename);
-
-  return 0;
 }
 
 unsigned long OpenTxEepromInterface::load(RadioData &radioData, const uint8_t * eeprom, int size)
@@ -733,15 +644,15 @@ int OpenTxFirmware::getCapability(Capability capability)
     case NumCurves:
       return (HAS_LARGE_LCD(board) ? 32 : (IS_ARM(board) ? 16 : 8));
     case HasMixerNames:
-      return (IS_ARM(board) ? (IS_TARANIS(board) ? 8 : 6) : false);
+      return (IS_ARM(board) ? (IS_HORUS_OR_TARANIS(board) ? 8 : 6) : false);
     case HasExpoNames:
-      return (IS_ARM(board) ? (IS_TARANIS(board) ? 8 : 6) : false);
+      return (IS_ARM(board) ? (IS_HORUS_OR_TARANIS(board) ? 8 : 6) : false);
     case HasNoExpo:
-      return (IS_TARANIS(board) ? false : true);
+      return (IS_HORUS_OR_TARANIS(board) ? false : true);
     case ChannelsName:
       return (IS_ARM(board) ? (HAS_LARGE_LCD(board) ? 6 : 4) : 0);
     case HasCvNames:
-      return (IS_TARANIS(board) ? 1 : 0);
+      return (IS_HORUS_OR_TARANIS(board) ? 1 : 0);
     case Telemetry:
       if (IS_2560(board) || IS_ARM(board) || id.contains("frsky") || id.contains("telemetrez"))
         return TM_HASTELEMETRY | TM_HASOFFSET | TM_HASWSHH;
@@ -796,6 +707,8 @@ int OpenTxFirmware::getCapability(Capability capability)
     case LcdWidth:
       if (IS_HORUS(board))
         return 480;
+      else if (IS_TARANIS_X7(board))
+        return 128;
       else if (IS_TARANIS(board))
         return 212;
       else
@@ -808,6 +721,8 @@ int OpenTxFirmware::getCapability(Capability capability)
     case LcdDepth:
       if (IS_HORUS(board))
         return 16;
+      else if (IS_TARANIS_X7(board))
+        return 1;
       else if (IS_TARANIS(board))
         return 4;
       else
@@ -1667,4 +1582,47 @@ void unregisterOpenTxFirmwares()
     foreach (Firmware * f, firmwares) {
       delete f;
     }
+}
+
+template <class T, class M>
+bool loadFromByteArray(T & dest, const QByteArray & data)
+{
+  foreach(OpenTxEepromInterface * eepromInterface, opentxEEpromInterfaces) {
+    if (eepromInterface->loadFromByteArray<T, M>(dest, data)) {
+      return true;
+    }
+  }
+  return false;
+}
+
+template <class T, class M>
+bool saveToByteArray(const T & dest, QByteArray & data)
+{
+  BoardEnum board = GetCurrentFirmware()->getBoard();
+  foreach(OpenTxEepromInterface * eepromInterface, opentxEEpromInterfaces) {
+    if (eepromInterface->getBoard() == board) {
+      return eepromInterface->saveToByteArray<T, M>(dest, data);
+    }
+  }
+  return false;
+}
+
+bool loadModelFromByteArray(ModelData & model, const QByteArray & data)
+{
+  return loadFromByteArray<ModelData, OpenTxModelData>(model, data);
+}
+
+bool loadRadioSettingsFromByteArray(GeneralSettings & settings, const QByteArray & data)
+{
+  return loadFromByteArray<GeneralSettings, OpenTxGeneralData>(settings, data);
+}
+
+bool writeModelToByteArray(const ModelData & model, QByteArray & data)
+{
+  return saveToByteArray<ModelData, OpenTxModelData>(model, data);
+}
+
+bool writeRadioSettingsToByteArray(const GeneralSettings & settings, QByteArray & data)
+{
+  return saveToByteArray<GeneralSettings, OpenTxGeneralData>(settings, data);
 }
