+++ companion/src/mainwindow.cpp
@@ -48,7 +48,6 @@
 #include "process_sync.h"
 #include "radiointerface.h"
 #include "progressdialog.h"
-#include "storage_sdcard.h"
 
 #define OPENTX_COMPANION_DOWNLOADS        "http://downloads-22.open-tx.org/companion"
 #define DONATE_STR                        "https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=QUZ48K4SEXDP2"
@@ -755,49 +754,13 @@ void MainWindow::loadBackup()
 
 void MainWindow::readEeprom()
 {
-  if(GetCurrentFirmware()->getBoard()== BOARD_HORUS && HORUS_READY_FOR_RELEASE()) {
-    // just an example
-    QString path = findMassstoragePath("RADIO");
-    if (path.isEmpty()) {
-      qDebug() << "Horus card not found";
-      return;
-    }
-
-    QString realPath = path.remove(path.size()- 5, 5);
-
-    qDebug() << "Reading files from" << realPath;
-    StorageSdcard storage;
-    storage.read(realPath);
-
-    // display models.txt
-    QString modelList = QString(storage.modelList);
-    qDebug() << "Models: size" << modelList.size() << "contents" << modelList;
-
-    // info about radio.bin
-    qDebug() << "Radio settings:" << storage.radio.size();
-
-    // info about all models
-    QList<QString> models = storage.getModelsFileNames();
-    qDebug() << "We have" << models.size() << "models:";
-    foreach(QString filename, models) {
-      QList<ModelFile>::const_iterator i = storage.getModelIterator(filename);
-      if (i != storage.models.end()) {
-        qDebug() << "\tModel:" << i->filename << "size" << i->data.size();
-      }
-    }
-
-    for (QList<ModelFile>::iterator i = storage.models.begin(); i != storage.models.end(); ++i) {
-    }
-
-
-    // for test immediately save to current dir
-    storage.write("./");
-
+  if (GetCurrentFirmware()->getBoard()== BOARD_HORUS && HORUS_READY_FOR_RELEASE()) {
+    // TODO
   }
   else {
     QString tempFile;
 
-    EEPROMInterface *eepromInterface = GetEepromInterface();
+    EEPROMInterface * eepromInterface = GetEepromInterface();
 
     if (IS_ARM(eepromInterface->getBoard()))
       tempFile = generateProcessUniqueTempFileName("temp.bin");
