+++ companion/src/mdichild.cpp
@@ -32,7 +32,7 @@
 #include "appdata.h"
 #include "wizarddialog.h"
 #include "flashfirmwaredialog.h"
-#include "storage_eeprom.h"
+#include "storage.h"
 
 #if defined _MSC_VER || !defined __GNUC__
 #include <windows.h>
@@ -386,147 +386,24 @@ void MdiChild::newFile()
   updateTitle();
 }
 
-bool MdiChild::loadFile(const QString & fileName, bool resetCurrentFile)
+bool MdiChild::loadFile(const QString & filename, bool resetCurrentFile)
 {
-  QFile file(fileName);
-
-  if (!file.exists()) {
-    QMessageBox::critical(this, tr("Error"), tr("Unable to find file %1!").arg(fileName));
+  Storage storage(filename);
+  if (!storage.load(radioData)) {
+    QMessageBox::critical(this, tr("Error"), storage.error());
     return false;
   }
 
-  int fileType = getFileType(fileName);
-
-#if 0
-  if (fileType==FILE_TYPE_XML) {
-    if (!file.open(QIODevice::ReadOnly | QIODevice::Text)) {  //reading HEX TEXT file
-      QMessageBox::critical(this, tr("Error"),
-          tr("Error opening file %1:\n%2.")
-          .arg(fileName)
-          .arg(file.errorString()));
-      return false;
-    }
-    QTextStream inputStream(&file);
-    XmlInterface(inputStream).load(radioData);
-  }
-  else
-#endif
-  if (fileType==FILE_TYPE_HEX || fileType==FILE_TYPE_EEPE) { //read HEX file
-    if (!file.open(QIODevice::ReadOnly | QIODevice::Text)) {  //reading HEX TEXT file
-        QMessageBox::critical(this, tr("Error"),
-                             tr("Error opening file %1:\n%2.")
-                             .arg(fileName)
-                             .arg(file.errorString()));
-        return false;
-    }
-
-    QDomDocument doc(ER9X_EEPROM_FILE_TYPE);
-    bool xmlOK = doc.setContent(&file);
-    if (xmlOK) {
-      std::bitset<NUM_ERRORS> errors((unsigned long long)LoadEepromXml(radioData, doc));
-      if (errors.test(ALL_OK)) {
-        refresh(true);
-        if (resetCurrentFile) setCurrentFile(fileName);
-        return true;
-      }
-    }
-    file.reset();
-
-    QTextStream inputStream(&file);
-
-    if (fileType==FILE_TYPE_EEPE) {  // read EEPE file header
-      QString hline = inputStream.readLine();
-      if (hline!=EEPE_EEPROM_FILE_HEADER) {
-        file.close();
-        return false;
-      }
-    }
-
-    QByteArray eeprom(EESIZE_MAX, 0);
-    int eeprom_size = HexInterface(inputStream).load((uint8_t *)eeprom.data(), EESIZE_MAX);
-    if (!eeprom_size) {
-      QMessageBox::critical(this, tr("Error"),
-          tr("Invalid EEPROM File %1")
-          .arg(fileName));
-      file.close();
-      return false;
-    }
-
-    file.close();
-
-    std::bitset<NUM_ERRORS> errors((unsigned long long)LoadEeprom(radioData, (uint8_t *)eeprom.data(), eeprom_size));
-    if (!errors.test(ALL_OK)) {
-      ShowEepromErrors(this, tr("Error"), tr("Invalid EEPROM File %1").arg(fileName), errors.to_ulong());
-      return false;
-    }
-    if (errors.test(HAS_WARNINGS)) {
-      ShowEepromWarnings(this, tr("Warning"), errors.to_ulong());
-    }
-
-    refresh(true);
-    if (resetCurrentFile) setCurrentFile(fileName);
-
-    return true;
-  }
-  else if (fileType==FILE_TYPE_BIN) { //read binary
-    int eeprom_size = file.size();
-
-    if (!file.open(QFile::ReadOnly)) {  //reading binary file   - TODO HEX support
-        QMessageBox::critical(this, tr("Error"),
-                             tr("Error opening file %1:\n%2.")
-                             .arg(fileName)
-                             .arg(file.errorString()));
-        return false;
-    }
-    uint8_t * eeprom = (uint8_t *)malloc(eeprom_size);
-    memset(eeprom, 0, eeprom_size);
-    long result = file.read((char*)eeprom, eeprom_size);
-    file.close();
-
-    if (result != eeprom_size) {
-        QMessageBox::critical(this, tr("Error"),
-                             tr("Error reading file %1:\n%2.")
-                             .arg(fileName)
-                             .arg(file.errorString()));
-        free(eeprom);
-        return false;
-    }
-
-    std::bitset<NUM_ERRORS> errorsEeprom((unsigned long long)LoadEeprom(radioData, eeprom, eeprom_size));
-    if (!errorsEeprom.test(ALL_OK)) {
-      std::bitset<NUM_ERRORS> errorsBackup((unsigned long long)LoadBackup(radioData, eeprom, eeprom_size, 0));
-      if (!errorsBackup.test(ALL_OK)) {
-        ShowEepromErrors(this, tr("Error"), tr("Invalid binary EEPROM File %1").arg(fileName), (errorsEeprom | errorsBackup).to_ulong());
-        free(eeprom);
-        return false;
-      }
-      if (errorsBackup.test(HAS_WARNINGS)) {
-        ShowEepromWarnings(this, tr("Warning"), errorsBackup.to_ulong());
-      }
-    }
-    else if (errorsEeprom.test(HAS_WARNINGS)) {
-      ShowEepromWarnings(this, tr("Warning"), errorsEeprom.to_ulong());
-    }
-
-    refresh(true);
-    if (resetCurrentFile)
-      setCurrentFile(fileName);
-
-    free(eeprom);
-    return true;
-  }
-  else if (fileType == FILE_TYPE_OTX) {
-    // read zip archive
-    bool loadFile(RadioData & radioData, const QString & filename);
-    if (loadFile(radioData, fileName)) {
-      refresh(true);
-      if (resetCurrentFile)
-        setCurrentFile(fileName);
-      return true;
-    }
+  QString warning = storage.warning();
+  if (!warning.isEmpty()) {
+    // TODO ShowEepromWarnings(this, tr("Warning"), warning);
   }
 
-  return false;
+  refresh(true);
+  if (resetCurrentFile)
+    setCurrentFile(filename);
+
+  return true;
 }
 
 bool MdiChild::save()
@@ -757,7 +634,7 @@ bool MdiChild::loadBackup()
     return false;
   }
   
-  int index = getCurrentRow();
+  // TODO int index = getCurrentRow();
 
   int eeprom_size = file.size();
   if (!file.open(QFile::ReadOnly)) {  //reading binary file   - TODO HEX support
@@ -779,7 +656,8 @@ bool MdiChild::loadBackup()
 
     return false;
   }
-
+  
+#if 0
   std::bitset<NUM_ERRORS> errorsEeprom((unsigned long long)LoadBackup(radioData, (uint8_t *)eeprom.data(), eeprom_size, index));
   if (!errorsEeprom.test(ALL_OK)) {
     ShowEepromErrors(this, tr("Error"), tr("Invalid binary backup File %1").arg(fileName), (errorsEeprom).to_ulong());
@@ -790,6 +668,8 @@ bool MdiChild::loadBackup()
   }
 
   refresh(true);
-
   return true;
+#else
+  return false;
+#endif
 }
