+++ companion/src/storage/sdcard.cpp
@@ -1,0 +1,59 @@
+/*
+ * Copyright (C) OpenTX
+ *
+ * Based on code named
+ *   th9x - http://code.google.com/p/th9x
+ *   er9x - http://code.google.com/p/er9x
+ *   gruvin9x - http://code.google.com/p/gruvin9x
+ *
+ * License GPLv2: http://www.gnu.org/licenses/gpl-2.0.html
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#include "sdcard.h"
+#include <QFile>
+#include <QDir>
+
+bool SdcardFormat::write(const RadioData & radioData)
+{
+  QDir dir(filename);
+  dir.mkdir("RADIO");
+  dir.mkdir("MODELS");
+  return CategorizedStorageFormat::write(radioData);
+}
+
+bool SdcardFormat::loadFile(QByteArray & filedata, const QString & filename)
+{
+  QString path = this->filename + "/" + filename;
+  QFile file(path);
+  if (!file.open(QFile::ReadOnly)) {
+    setError(QObject::tr("Error opening file %1:\n%2.").arg(path).arg(file.errorString()));
+    return false;
+  }
+  filedata = file.readAll();
+  qDebug() << "File" << path << "read, size:" << filedata.size();
+  return true;
+}
+
+bool SdcardFormat::writeFile(const QByteArray & data, const QString & filename)
+{
+  QString path = this->filename + "/" + filename;
+  QFile file(path);
+  if (!file.open(QFile::WriteOnly)) {
+    setError(QObject::tr("Error opening file %1:\n%2.").arg(path).arg(file.errorString()));
+    qDebug() << "File" << path << "write error";
+    return false;
+  }
+  file.write(data.data(), data.size());
+  file.close();
+  qDebug() << "File" << path << "written, size:" << data.size();
+  return true;
+}
